requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
WTForms==3.0.1
SQLite3==3.37.0
goprocam==4.2.0
requests==2.27.1
opencv-python==4.5.5.62
numpy==1.22.1
python-dotenv==0.19.2